Actions
Arquivo cycles.ts foi modificado, nome foi trocado para reducer.ts

Foi criado uma pasta cycles, para compor o componente.

Criado a tipificação de nossas ações, no componente action foi criado funções para realizar o controle dos ciclos, alem de
